Move server related files into server directory for clarity
