+++ .gitignore
@@ -1,3 +1,2 @@
 node_modules
-creds.txt
-seeder/config.js
+server/creds.txt
