--- STYLE-GUIDE.md
@@ -1,4 +1,0 @@
-## Style Guide
-
-Refer to the [AirBnb Style Guide](https://github.com/airbnb/javascript).
-
