+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "build": "webpack --config webpack.config.js --watch",
     "start": "nodemon server/index.js",
-    "seed": "mysql --defaults-extra-file=creds.txt < schema.sql && node seed.js"
+    "seed": "mysql --defaults-extra-file=server/creds.txt < server/schema.sql && node server/seed.js"
   },
   "dependencies": {
     "express": "^4.16.4",
@@ -24,7 +24,9 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.5",
+    "css-loader": "^2.1.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "style-loader": "^0.23.1",
     "webpack": "^4.29.3",
     "webpack-cli": "^3.2.3"
   },
