+++ public/app.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import Reservation from './reservation.js';
 
 class App extends React.Component {
   constructor (props) {
@@ -14,7 +14,7 @@ class App extends React.Component {
 
   render () {
     return (
-      <div>TESTERINO!</div>
+      <Reservation />
     )
   }
 }
