+++ public/reservation.js
@@ -1,0 +1,29 @@
+import React from 'react';
+
+class Reservation extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+    }
+  }
+
+  render () {
+    return (
+      <div>
+        <div>
+          <h3>
+            <span>Icon</span>
+            <a>Make a Reservation</a>
+          </h3>
+        </div>
+
+        <div>
+          <select>
+          </select>
+        </div>
+      </div>
+    )
+  }
+}
+
+export default Reservation;
