+++ public/styles/app.css
@@ -1,0 +1,3 @@
+div {
+  background-color: green;
+}
